Added the type double for square

+++ src/Lib.hs
@@ -3,5 +3,5 @@ module Lib (someFunc, square) where
 someFunc :: IO ()
 someFunc = putStrLn "someFunc"
 
-square :: Integer -> Integer
+square :: Double -> Double
 square n = n * n
